backend/requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
mysql-connector-python